Change timeout_multiplier in platform specification to float

+++ src/twister2/platform_specification.py
@@ -19,7 +19,7 @@ class Testing:
     default: bool = False
     only_tags: set[str] = field(default_factory=set)
     ignore_tags: set[str] = field(default_factory=set)
-    timeout_multiplier: int = 1
+    timeout_multiplier: float = 1
 
     def __post_init__(self):
         self.only_tags = string_to_set(self.only_tags)
@@ -85,7 +85,7 @@ class TestingSchema(Schema):
     default = fields.Bool()
     only_tags = fields.List(fields.Str)
     ignore_tags = fields.List(fields.Str)
-    timeout_multiplier = fields.Int()
+    timeout_multiplier = fields.Float()
 
 
 class PlatformSchema(Schema):
